refactor(backend): modify RelationName import bb-166
